requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
streamlit>=1.28.0